+++ extensions/itranslate/CHANGELOG.md
@@ -1,5 +1,10 @@
 # iTranslate Changelog
 
+## [Translation History & TTS] - 2022-09-02
+
+- Add new command `Translation History`. Support record translation history
+- Support TTS by OS command `say`
+
 ## [Add New Translation Service Providers & New icon] - 2022-08-26
 
 - Add Google Could Translation option